+++ client/src/client/client_future.rs
@@ -6,6 +6,7 @@
 // copied, modified, or distributed except according to those terms.
 
 use std::collections::{HashMap, HashSet};
+use std::marker::PhantomData;
 use std::io;
 use std::time::Duration;
 
@@ -52,16 +53,17 @@ impl ClientStreamHandle for StreamHandle {
 ///  implementations.
 #[must_use = "futures do nothing unless polled"]
 pub struct ClientFuture<S: Stream<Item = Vec<u8>, Error = io::Error>> {
-    stream: S,
-    reactor_handle: Handle,
-    timeout_duration: Duration,
-    // TODO genericize and remove this Box
-    stream_handle: Box<ClientStreamHandle>,
-    new_receiver:
-        Peekable<StreamFuse<UnboundedReceiver<(Message, Complete<ClientResult<Message>>)>>>,
-    active_requests: HashMap<u16, (Complete<ClientResult<Message>>, Timeout)>,
-    // TODO: Maybe make a typed version of ClientFuture for Updates?
-    signer: Option<Signer>,
+    phantom: PhantomData<S>,
+    // stream: S,
+    // reactor_handle: Handle,
+    // timeout_duration: Duration,
+    // // TODO genericize and remove this Box
+    // stream_handle: Box<ClientStreamHandle>,
+    // new_receiver:
+    //     Peekable<StreamFuse<UnboundedReceiver<(Message, Complete<ClientResult<Message>>)>>>,
+    // active_requests: HashMap<u16, (Complete<ClientResult<Message>>, Timeout)>,
+    // // TODO: Maybe make a typed version of ClientFuture for Updates?
+    // signer: Option<Signer>,
 }
 
 impl<S: Stream<Item = Vec<u8>, Error = io::Error> + 'static> ClientFuture<S> {
@@ -103,16 +105,21 @@ impl<S: Stream<Item = Vec<u8>, Error = io::Error> + 'static> ClientFuture<S> {
     /// * `timeout_duration` - All requests may fail due to lack of response, this is the time to
     ///                        wait for a response before canceling the request.
     /// * `stream_handle` - The handle for the `stream` on which bytes can be sent/received.
-    /// * `signer` - An optional signer for requests, needed for Updates with Sig0, otherwise not needed
+    /// * `finalizer` - An optional signer for requests, needed for Updates with Sig0, otherwise not needed
     pub fn with_timeout(
         stream: Box<Future<Item = S, Error = io::Error>>,
         stream_handle: Box<DnsStreamHandle>,
         loop_handle: &Handle,
         timeout_duration: Duration,
-        signer: Option<Signer>,
+        finalizer: Option<Signer>,
     ) -> BasicClientHandle {
-        let dns_future =
-            DnsFuture::with_timeout(stream, stream_handle, loop_handle, timeout_duration, signer);
+        let dns_future_handle = DnsFuture::with_timeout(
+            stream,
+            stream_handle,
+            loop_handle,
+            timeout_duration,
+            finalizer,
+        );
 
         // let (sender, rx) = unbounded();
 
@@ -143,256 +150,256 @@ impl<S: Stream<Item = Vec<u8>, Error = io::Error> + 'static> ClientFuture<S> {
         //         }),
         // );
 
-        BasicClientHandle { message_sender: dns_future }
+        BasicClientHandle { message_sender: dns_future_handle }
     }
 
-    /// loop over active_requests and remove cancelled requests
-    ///  this should free up space if we already had 4096 active requests
-    fn drop_cancelled(&mut self) {
-        // TODO: should we have a timeout here? or always expect the caller to do this?
-        let mut canceled = HashSet::new();
-        for (&id, &mut (ref mut req, ref mut timeout)) in self.active_requests.iter_mut() {
-            if let Ok(Async::Ready(())) = req.poll_cancel() {
-              canceled.insert(id);
-            }
+    // /// loop over active_requests and remove cancelled requests
+    // ///  this should free up space if we already had 4096 active requests
+    // fn drop_cancelled(&mut self) {
+    //     // TODO: should we have a timeout here? or always expect the caller to do this?
+    //     let mut canceled = HashSet::new();
+    //     for (&id, &mut (ref mut req, ref mut timeout)) in self.active_requests.iter_mut() {
+    //         if let Ok(Async::Ready(())) = req.poll_cancel() {
+    //           canceled.insert(id);
+    //         }
 
-            // check for timeouts...
-            match timeout.poll() {
-                Ok(Async::Ready(_)) => {
-                    warn!("request timeout: {}", id);
-                    canceled.insert(id);
-                }
-                Ok(Async::NotReady) => (),
-                Err(e) => {
-                    error!("unexpected error from timeout: {}", e);
-                    canceled.insert(id);
-                }
-            }
-        }
+    //         // check for timeouts...
+    //         match timeout.poll() {
+    //             Ok(Async::Ready(_)) => {
+    //                 warn!("request timeout: {}", id);
+    //                 canceled.insert(id);
+    //             }
+    //             Ok(Async::NotReady) => (),
+    //             Err(e) => {
+    //                 error!("unexpected error from timeout: {}", e);
+    //                 canceled.insert(id);
+    //             }
+    //         }
+    //     }
 
-        // drop all the canceled requests
-        for id in canceled {
-            if let Some((req, _)) = self.active_requests.remove(&id) {
-                // TODO, perhaps there is a different reason timeout? but there shouldn't be...
-                //  being lazy and always returning timeout in this case (if it was canceled then the
-                //  then the otherside isn't really paying attention anyway)
+    //     // drop all the canceled requests
+    //     for id in canceled {
+    //         if let Some((req, _)) = self.active_requests.remove(&id) {
+    //             // TODO, perhaps there is a different reason timeout? but there shouldn't be...
+    //             //  being lazy and always returning timeout in this case (if it was canceled then the
+    //             //  then the otherside isn't really paying attention anyway)
 
-                // complete the request, it's failed...
-                req.send(Err(ClientErrorKind::Timeout.into())).expect(
-                    "error notifying wait, possible future leak",
-                );
-            }
-        }
-    }
+    //             // complete the request, it's failed...
+    //             req.send(Err(ClientErrorKind::Timeout.into())).expect(
+    //                 "error notifying wait, possible future leak",
+    //             );
+    //         }
+    //     }
+    // }
 
-    /// creates random query_id, validates against all active queries
-    fn next_random_query_id(&self) -> Async<u16> {
-        let mut rand = rand::thread_rng();
+    // /// creates random query_id, validates against all active queries
+    // fn next_random_query_id(&self) -> Async<u16> {
+    //     let mut rand = rand::thread_rng();
 
-        for _ in 0..100 {
-            let id = rand.gen_range(0_u16, u16::max_value());
+    //     for _ in 0..100 {
+    //         let id = rand.gen_range(0_u16, u16::max_value());
 
-            if !self.active_requests.contains_key(&id) {
-                return Async::Ready(id);
-            }
-        }
+    //         if !self.active_requests.contains_key(&id) {
+    //             return Async::Ready(id);
+    //         }
+    //     }
 
-        warn!("could not get next random query id, delaying");
-        task::current().notify();
-        Async::NotReady
-    }
+    //     warn!("could not get next random query id, delaying");
+    //     task::current().notify();
+    //     Async::NotReady
+    // }
 }
 
-impl<S: Stream<Item = Vec<u8>, Error = io::Error> + 'static> Future for ClientFuture<S> {
-    type Item = ();
-    type Error = ClientError;
+// impl<S: Stream<Item = Vec<u8>, Error = io::Error> + 'static> Future for ClientFuture<S> {
+//     type Item = ();
+//     type Error = ClientError;
 
-    fn poll(&mut self) -> Poll<(), Self::Error> {
-        self.drop_cancelled();
+//     fn poll(&mut self) -> Poll<(), Self::Error> {
+//         self.drop_cancelled();
 
-        // loop over new_receiver for all outbound requests
-        loop {
-            // get next query_id
-            let query_id: Option<u16> = match self.new_receiver.peek() {
-                Ok(Async::Ready(Some(_))) => {
-                    debug!("got message from receiver");
+//         // loop over new_receiver for all outbound requests
+//         loop {
+//             // get next query_id
+//             let query_id: Option<u16> = match self.new_receiver.peek() {
+//                 Ok(Async::Ready(Some(_))) => {
+//                     debug!("got message from receiver");
 
-                    // we have a new message to send
-                    match self.next_random_query_id() {
-                        Async::Ready(id) => Some(id),
-                        Async::NotReady => break,
-                    }
-                }
-                Ok(_) => None,
-                Err(()) => {
-          warn!("receiver was shutdown?");
-          break
-        }
-            };
+//                     // we have a new message to send
+//                     match self.next_random_query_id() {
+//                         Async::Ready(id) => Some(id),
+//                         Async::NotReady => break,
+//                     }
+//                 }
+//                 Ok(_) => None,
+//                 Err(()) => {
+//           warn!("receiver was shutdown?");
+//           break
+//         }
+//             };
 
-            // finally pop the reciever
-            match self.new_receiver.poll() {
-                Ok(Async::Ready(Some((mut message, complete)))) => {
-                    // if there was a message, and the above succesion was succesful,
-                    //  register the new message, if not do not register, and set the complete to error.
-                    // getting a random query id, this mitigates potential cache poisoning.
-                    // TODO: for SIG0 we can't change the message id after signing.
-                    let query_id = query_id.expect("query_id should have been set above");
-                    message.set_id(query_id);
+//             // finally pop the reciever
+//             match self.new_receiver.poll() {
+//                 Ok(Async::Ready(Some((mut message, complete)))) => {
+//                     // if there was a message, and the above succesion was succesful,
+//                     //  register the new message, if not do not register, and set the complete to error.
+//                     // getting a random query id, this mitigates potential cache poisoning.
+//                     // TODO: for SIG0 we can't change the message id after signing.
+//                     let query_id = query_id.expect("query_id should have been set above");
+//                     message.set_id(query_id);
 
-                    // update messages need to be signed.
-                    if let OpCode::Update = message.op_code() {
-                        if let Some(ref signer) = self.signer {
-                            // TODO: it's too bad this happens here...
-                            if let Err(e) = message.sign(signer, Utc::now().timestamp() as u32) {
-                                warn!("could not sign message: {}", e);
-                                complete.send(Err(e.into())).expect(
-                                    "error notifying wait, possible future leak",
-                                );
-                                continue; // to the next message...
-                            }
-                        }
-                    }
+//                     // update messages need to be signed.
+//                     if let OpCode::Update = message.op_code() {
+//                         if let Some(ref signer) = self.signer {
+//                             // TODO: it's too bad this happens here...
+//                             if let Err(e) = message.sign(signer, Utc::now().timestamp() as u32) {
+//                                 warn!("could not sign message: {}", e);
+//                                 complete.send(Err(e.into())).expect(
+//                                     "error notifying wait, possible future leak",
+//                                 );
+//                                 continue; // to the next message...
+//                             }
+//                         }
+//                     }
 
-                    // store a Timeout for this message before sending
-                    let timeout = match Timeout::new(self.timeout_duration, &self.reactor_handle) {
-                        Ok(timeout) => timeout,
-                        Err(e) => {
-                            warn!("could not create timer: {}", e);
-                            complete.send(Err(e.into())).expect(
-                                "error notifying wait, possible future leak",
-                            );
-                            continue; // to the next message...
-                        }
-                    };
+//                     // store a Timeout for this message before sending
+//                     let timeout = match Timeout::new(self.timeout_duration, &self.reactor_handle) {
+//                         Ok(timeout) => timeout,
+//                         Err(e) => {
+//                             warn!("could not create timer: {}", e);
+//                             complete.send(Err(e.into())).expect(
+//                                 "error notifying wait, possible future leak",
+//                             );
+//                             continue; // to the next message...
+//                         }
+//                     };
 
-                    // send the message
-                    match message.to_vec() {
-                        Ok(buffer) => {
-                            debug!("sending message id: {}", query_id);
-                            try!(self.stream_handle.send(buffer));
-                            // add to the map -after- the client send b/c we don't want to put it in the map if
-                            //  we ended up returning from the send.
-                            self.active_requests.insert(
-                                message.id(),
-                                (complete, timeout),
-                            );
-                        }
-                        Err(e) => {
-                            debug!("error message id: {} error: {}", query_id, e);
-                            // complete with the error, don't add to the map of active requests
-                            complete.send(Err(e.into())).expect(
-                                "error notifying wait, possible future leak",
-                            );
-                        }
-                    }
-                }
-                Ok(_) => break,
-                Err(()) => {
-          warn!("receiver was shutdown?");
-          break
-        }
-            }
-        }
+//                     // send the message
+//                     match message.to_vec() {
+//                         Ok(buffer) => {
+//                             debug!("sending message id: {}", query_id);
+//                             try!(self.stream_handle.send(buffer));
+//                             // add to the map -after- the client send b/c we don't want to put it in the map if
+//                             //  we ended up returning from the send.
+//                             self.active_requests.insert(
+//                                 message.id(),
+//                                 (complete, timeout),
+//                             );
+//                         }
+//                         Err(e) => {
+//                             debug!("error message id: {} error: {}", query_id, e);
+//                             // complete with the error, don't add to the map of active requests
+//                             complete.send(Err(e.into())).expect(
+//                                 "error notifying wait, possible future leak",
+//                             );
+//                         }
+//                     }
+//                 }
+//                 Ok(_) => break,
+//                 Err(()) => {
+//           warn!("receiver was shutdown?");
+//           break
+//         }
+//             }
+//         }
 
-        // Collect all inbound requests, max 100 at a time for QoS
-        //   by having a max we will guarantee that the client can't be DOSed in this loop
-        // TODO: make the QoS configurable
-        let mut messages_received = 0;
-        for i in 0..QOS_MAX_RECEIVE_MSGS {
-            match try!(self.stream.poll()) {
-                Async::Ready(Some(buffer)) => {
-                    messages_received = i;
+//         // Collect all inbound requests, max 100 at a time for QoS
+//         //   by having a max we will guarantee that the client can't be DOSed in this loop
+//         // TODO: make the QoS configurable
+//         let mut messages_received = 0;
+//         for i in 0..QOS_MAX_RECEIVE_MSGS {
+//             match try!(self.stream.poll()) {
+//                 Async::Ready(Some(buffer)) => {
+//                     messages_received = i;
 
-                    //   deserialize or log decode_error
-                    match Message::from_vec(&buffer) {
-                        Ok(message) => {
-                            match self.active_requests.remove(&message.id()) {
-                                Some((complete, _)) => {
-                                    complete.send(Ok(message)).expect(
-                                        "error notifying wait, possible future leak",
-                                    )
-                                }
-                                None => debug!("unexpected request_id: {}", message.id()),
-                            }
-                        }
-                        // TODO: return src address for diagnostics
-                        Err(e) => debug!("error decoding message: {}", e),
-                    }
+//                     //   deserialize or log decode_error
+//                     match Message::from_vec(&buffer) {
+//                         Ok(message) => {
+//                             match self.active_requests.remove(&message.id()) {
+//                                 Some((complete, _)) => {
+//                                     complete.send(Ok(message)).expect(
+//                                         "error notifying wait, possible future leak",
+//                                     )
+//                                 }
+//                                 None => debug!("unexpected request_id: {}", message.id()),
+//                             }
+//                         }
+//                         // TODO: return src address for diagnostics
+//                         Err(e) => debug!("error decoding message: {}", e),
+//                     }
 
-                }
-                Async::Ready(None) |
-                Async::NotReady => break,
-            }
-        }
+//                 }
+//                 Async::Ready(None) |
+//                 Async::NotReady => break,
+//             }
+//         }
 
-        // Clean shutdown happens when all pending requests are done and the
-        // incoming channel has been closed (e.g. you'll never receive another
-        // request). try! will early return the error...
-        let done = if let Async::Ready(None) = try!(self.new_receiver.peek()) {
-            true
-        } else {
-            false
-        };
-        if self.active_requests.is_empty() && done {
-            return Ok(().into()); // we are done
-        }
+//         // Clean shutdown happens when all pending requests are done and the
+//         // incoming channel has been closed (e.g. you'll never receive another
+//         // request). try! will early return the error...
+//         let done = if let Async::Ready(None) = try!(self.new_receiver.peek()) {
+//             true
+//         } else {
+//             false
+//         };
+//         if self.active_requests.is_empty() && done {
+//             return Ok(().into()); // we are done
+//         }
 
-        // If still active, then if the qos (for _ in 0..100 loop) limit
-        // was hit then "yield". This'll make sure that the future is
-        // woken up immediately on the next turn of the event loop.
-        if messages_received == QOS_MAX_RECEIVE_MSGS {
-            task::current().notify();
-        }
+//         // If still active, then if the qos (for _ in 0..100 loop) limit
+//         // was hit then "yield". This'll make sure that the future is
+//         // woken up immediately on the next turn of the event loop.
+//         if messages_received == QOS_MAX_RECEIVE_MSGS {
+//             task::current().notify();
+//         }
 
-        // Finally, return not ready to keep the 'driver task' alive.
-        return Ok(Async::NotReady);
-    }
-}
+//         // Finally, return not ready to keep the 'driver task' alive.
+//         return Ok(Async::NotReady);
+//     }
+// }
 
-/// Always returns the specified io::Error to the remote Sender
-struct ClientStreamErrored {
-    error: io::Error,
-    new_receiver:
-        Peekable<StreamFuse<UnboundedReceiver<(Message, Complete<ClientResult<Message>>)>>>,
-}
+// /// Always returns the specified io::Error to the remote Sender
+// struct ClientStreamErrored {
+//     error: io::Error,
+//     new_receiver:
+//         Peekable<StreamFuse<UnboundedReceiver<(Message, Complete<ClientResult<Message>>)>>>,
+// }
 
-impl Future for ClientStreamErrored {
-    type Item = ();
-    type Error = ClientError;
+// impl Future for ClientStreamErrored {
+//     type Item = ();
+//     type Error = ClientError;
 
-    fn poll(&mut self) -> Poll<(), Self::Error> {
-        match self.new_receiver.poll() {
-            Ok(Async::Ready(Some((_, complete)))) => {
-                complete
-                    .send(Err(ClientError::from(&self.error).clone()))
-                    .expect("error notifying wait, possible future leak");
+//     fn poll(&mut self) -> Poll<(), Self::Error> {
+//         match self.new_receiver.poll() {
+//             Ok(Async::Ready(Some((_, complete)))) => {
+//                 complete
+//                     .send(Err(ClientError::from(&self.error).clone()))
+//                     .expect("error notifying wait, possible future leak");
 
-                task::current().notify();
-                return Ok(Async::NotReady);
-            }
-            Ok(Async::Ready(None)) => return Ok(Async::Ready(())),            
-            _ => return Err(ClientErrorKind::NoError.into()),
-        }
-    }
-}
+//                 task::current().notify();
+//                 return Ok(Async::NotReady);
+//             }
+//             Ok(Async::Ready(None)) => return Ok(Async::Ready(())),
+//             _ => return Err(ClientErrorKind::NoError.into()),
+//         }
+//     }
+// }
 
-enum ClientStreamOrError<S: Stream<Item = Vec<u8>, Error = io::Error> + 'static> {
-    Future(ClientFuture<S>),
-    Errored(ClientStreamErrored),
-}
+// enum ClientStreamOrError<S: Stream<Item = Vec<u8>, Error = io::Error> + 'static> {
+//     Future(ClientFuture<S>),
+//     Errored(ClientStreamErrored),
+// }
 
-impl<S: Stream<Item = Vec<u8>, Error = io::Error> + 'static> Future for ClientStreamOrError<S> {
-    type Item = ();
-    type Error = ClientError;
+// impl<S: Stream<Item = Vec<u8>, Error = io::Error> + 'static> Future for ClientStreamOrError<S> {
+//     type Item = ();
+//     type Error = ClientError;
 
-    fn poll(&mut self) -> Poll<(), Self::Error> {
-        match *self {
-            ClientStreamOrError::Future(ref mut f) => f.poll(),
-            ClientStreamOrError::Errored(ref mut e) => e.poll(),
-        }
-    }
-}
+//     fn poll(&mut self) -> Poll<(), Self::Error> {
+//         match *self {
+//             ClientStreamOrError::Future(ref mut f) => f.poll(),
+//             ClientStreamOrError::Errored(ref mut e) => e.poll(),
+//         }
+//     }
+// }
 
 
 /// Root ClientHandle implementaton returned by ClientFuture
@@ -400,6 +407,7 @@ impl<S: Stream<Item = Vec<u8>, Error = io::Error> + 'static> Future for ClientSt
 /// This can be used directly to perform queries. See `trust_dns::client::SecureClientHandle` for
 ///  a DNSSEc chain validator.
 #[derive(Clone)]
+#[deprecated(note = "See [`BasicDnsHandle`]")]
 pub struct BasicClientHandle {
     message_sender: BasicDnsHandle,
 }